Filter bug for null value #117
