+++ src/main/java/org/cboard/dataprovider/aggregator/Aggregatable.java
@@ -1,9 +1,14 @@
 package org.cboard.dataprovider.aggregator;
 
+import org.cboard.dataprovider.DataProvider;
 import org.cboard.dataprovider.config.AggConfig;
+import org.cboard.dataprovider.config.CompositeConfig;
+import org.cboard.dataprovider.config.ConfigComponent;
+import org.cboard.dataprovider.config.DimensionConfig;
 import org.cboard.dataprovider.result.AggregateResult;
 
-import java.util.Map;
+import java.util.ArrayList;
+import java.util.stream.Collectors;
 
 /**
  * Created by yfyuan on 2017/1/13.
@@ -35,5 +40,26 @@ public interface Aggregatable {
 
     default String viewAggDataQuery(AggConfig ac) throws Exception {
         return "Not Support";
+    }
+
+    default ConfigComponent separateNull(ConfigComponent configComponent) {
+        if (configComponent instanceof DimensionConfig) {
+            DimensionConfig cc = (DimensionConfig) configComponent;
+            if (("=".equals(cc.getFilterType()) || "≠".equals(cc.getFilterType())) && cc.getValues().size() > 1 &&
+                    cc.getValues().stream().anyMatch(s -> DataProvider.NULL_STRING.equals(s))) {
+                CompositeConfig compositeConfig = new CompositeConfig();
+                compositeConfig.setType("=".equals(cc.getFilterType()) ? "OR" : "AND");
+                cc.setValues(cc.getValues().stream().filter(s -> !DataProvider.NULL_STRING.equals(s)).collect(Collectors.toList()));
+                compositeConfig.getConfigComponents().add(cc);
+                DimensionConfig nullCc = new DimensionConfig();
+                nullCc.setColumnName(cc.getColumnName());
+                nullCc.setFilterType(cc.getFilterType());
+                nullCc.setValues(new ArrayList<>());
+                nullCc.getValues().add(DataProvider.NULL_STRING);
+                compositeConfig.getConfigComponents().add(nullCc);
+                return compositeConfig;
+            }
+        }
+        return configComponent;
     }
 }
