+++ src/main/java/org/cboard/dataprovider/aggregator/jvm/JvmAggregator.java
@@ -3,9 +3,15 @@ package org.cboard.dataprovider.aggregator.jvm;
 import com.alibaba.fastjson.JSONObject;
 import com.google.common.base.Charsets;
 import com.google.common.hash.Hashing;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.math.NumberUtils;
 import org.cboard.cache.CacheManager;
+import org.cboard.dataprovider.DataProvider;
 import org.cboard.dataprovider.aggregator.InnerAggregator;
 import org.cboard.dataprovider.config.AggConfig;
+import org.cboard.dataprovider.config.CompositeConfig;
+import org.cboard.dataprovider.config.ConfigComponent;
+import org.cboard.dataprovider.config.DimensionConfig;
 import org.cboard.dataprovider.result.AggregateResult;
 import org.cboard.dataprovider.result.ColumnIndex;
 import org.cboard.util.NaturalOrderComparator;
@@ -16,10 +22,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Service;
 
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
@@ -58,12 +61,10 @@ public class JvmAggregator extends InnerAggregator {
         Map<String, Integer> columnIndex = getColumnIndex(data);
         final int fi = columnIndex.get(columnName);
         Filter rowFilter = new Filter(config, columnIndex);
-        NaturalOrderComparator comparator = new NaturalOrderComparator();
         String[] filtered = Arrays.stream(data).parallel().skip(1)
                 .filter(rowFilter::filter)
                 .map(e -> e[fi])
                 .distinct()
-                .sorted(comparator)
                 .toArray(String[]::new);
         return filtered;
     }
@@ -133,6 +134,110 @@ public class JvmAggregator extends InnerAggregator {
         public int hashCode() {
             return Arrays.hashCode(dimensions);
         }
+    }
+
+    private class Filter {
+
+        private List<ConfigComponent> ruleList;
+        private Map<String, Integer> columnIndex;
+        private Comparator comparator = new NaturalOrderComparator();
+
+        public Filter(AggConfig config, Map<String, Integer> columnIndex) {
+            ruleList = new ArrayList<>();
+            if (config != null) {
+                ruleList.addAll(config.getColumns());
+                ruleList.addAll(config.getRows());
+                ruleList.addAll(config.getFilters());
+            }
+            this.columnIndex = columnIndex;
+        }
+
+        private boolean checkConfigComponent(ConfigComponent component, String[] row) {
+            if (component instanceof DimensionConfig) {
+                return checkRule((DimensionConfig) component, row);
+            } else if (component instanceof CompositeConfig) {
+                CompositeConfig compositeConfig = (CompositeConfig) component;
+                if (compositeConfig.getConfigComponents().size() < 1) {
+                    return false;
+                }
+                if ("AND".equalsIgnoreCase(compositeConfig.getType())) {
+                    return compositeConfig.getConfigComponents().stream().allMatch(e -> checkConfigComponent(e, row));
+                } else if ("OR".equalsIgnoreCase(compositeConfig.getType())) {
+                    return compositeConfig.getConfigComponents().stream().anyMatch(e -> checkConfigComponent(e, row));
+                }
+            }
+            return false;
+        }
+
+        private boolean checkRule(DimensionConfig rule, String[] row) {
+            if (rule.getValues().size() == 0 || rule.getValues().get(0) == null) {
+                return true;
+            }
+            String v = row[columnIndex.get(rule.getColumnName())];
+            String a = rule.getValues().get(0);
+            String b = rule.getValues().size() >= 2 ? rule.getValues().get(1) : null;
+            if (DataProvider.NULL_STRING.equals(a)) {
+                switch ((rule.getFilterType())) {
+                    case "=":
+                        return v == null;
+                    case "≠":
+                        return v != null;
+                }
+            }
+            if (StringUtils.isEmpty(v)) {
+                return false;
+            }
+            switch (rule.getFilterType()) {
+                case "=":
+                case "eq":
+                    return rule.getValues().stream().anyMatch(e -> e.equals(v));
+                case "≠":
+                case "ne":
+                    return rule.getValues().stream().allMatch(e -> !e.equals(v));
+                case ">":
+                    return comparator.compare(v, a) > 0;
+                case "<":
+                    return comparator.compare(v, a) < 0;
+                case "≥":
+                    return comparator.compare(v, a) >= 0;
+                case "≤":
+                    return comparator.compare(v, a) <= 0;
+                case "(a,b]":
+                    return (rule.getValues().size() >= 2) &&
+                            (comparator.compare(v, a) > 0) &&
+                            (comparator.compare(v, b) <= 0);
+                case "[a,b)":
+                    return (rule.getValues().size() >= 2) &&
+                            (comparator.compare(v, a) >= 0) &&
+                            (comparator.compare(v, b) < 0);
+                case "(a,b)":
+                    return (rule.getValues().size() >= 2) &&
+                            (comparator.compare(v, a) > 0) &&
+                            (comparator.compare(v, b) < 0);
+                case "[a,b]":
+                    return (rule.getValues().size() >= 2) &&
+                            (comparator.compare(v, a) >= 0) &&
+                            (comparator.compare(v, b) <= 0);
+            }
+            return true;
+        }
+
+        public boolean filter(String[] row) {
+            boolean result = ruleList.stream().map(rule -> separateNull(rule)).allMatch(rule ->
+                    checkConfigComponent(rule, row)
+            );
+            return result;
+        }
+
+        private Comparable[] tryToNumber(String... args) {
+            boolean allNumber = Arrays.stream(args).allMatch(e -> NumberUtils.isNumber(e));
+            if (allNumber) {
+                return Arrays.stream(args).mapToDouble(Double::parseDouble).boxed().toArray(Double[]::new);
+            } else {
+                return args;
+            }
+        }
+
     }
 }
